refactor: unify native backend via HAL abstraction
Eliminate ~950 lines of code duplication between vulkan.go and metal.go.
Single backend.go implementation uses hal.Device/hal.Queue interfaces,
with thin platform files (hal_vulkan.go, hal_metal.go) for backend selection.
Metal now gets all compute/buffer/fence operations through HAL.
