+++ gpu/backend/native/backend.go
@@ -1,38 +1,43 @@
-//go:build darwin
+//go:build windows || linux || darwin
 
-// Package gpu provides the WebGPU backend using pure Go (gogpu/wgpu).
+// Package native provides the WebGPU backend using pure Go (gogpu/wgpu).
 // This backend offers zero dependencies and simple cross-compilation.
 //
-// Implementation uses gogpu/wgpu HAL (Hardware Abstraction Layer) with Metal backend.
+// Implementation uses gogpu/wgpu HAL (Hardware Abstraction Layer) with
+// platform-specific backends: Vulkan on Windows/Linux, Metal on macOS.
 package native
 
 import (
+	"errors"
 	"fmt"
+	"sort"
+	"time"
 
 	"github.com/gogpu/gogpu/gpu"
 	"github.com/gogpu/gogpu/gpu/types"
 	"github.com/gogpu/gputypes"
 	"github.com/gogpu/wgpu/hal"
-	"github.com/gogpu/wgpu/hal/metal"
 )
 
 // Backend implements gpu.Backend using pure Go wgpu HAL.
 type Backend struct {
-	registry *ResourceRegistry
-	backend  hal.Backend
+	registry    *ResourceRegistry
+	backend     hal.Backend
+	bufferSizes map[types.Buffer]uint64
 }
 
 // New creates a new Pure Go backend.
 func New() *Backend {
 	return &Backend{
-		registry: NewResourceRegistry(),
-		backend:  metal.Backend{}, // Metal is the HAL implementation for macOS
+		registry:    NewResourceRegistry(),
+		backend:     newHalBackend(),
+		bufferSizes: make(map[types.Buffer]uint64),
 	}
 }
 
 // Name returns the backend identifier.
 func (b *Backend) Name() string {
-	return "Pure Go (gogpu/wgpu/metal)"
+	return halBackendName()
 }
 
 // Init initializes the backend.
@@ -44,9 +49,11 @@ func (b *Backend) Init() error {
 
 // Destroy releases all backend resources.
 func (b *Backend) Destroy() {
-	// Note: This does NOT destroy HAL resources!
-	// Caller must explicitly release all handles before calling Destroy.
-	// This just clears the registry.
+	// Wait for all GPU operations to complete before destroying resources.
+	// This prevents hangs/crashes when closing the window.
+	b.registry.WaitAllDevicesIdle()
+
+	// Clear the registry (does not destroy HAL resources, but they will be GC'd)
 	b.registry.Clear()
 }
 
@@ -54,8 +61,8 @@ func (b *Backend) Destroy() {
 func (b *Backend) CreateInstance() (types.Instance, error) {
 	// Create HAL instance with default config
 	desc := &hal.InstanceDescriptor{
-		Backends: gputypes.Backends(1 << gputypes.BackendMetal), // Metal backend
-		Flags:    0,                                             // No debug for now
+		Backends: gputypes.Backends(1 << halBackendVariant()),
+		Flags:    0, // No debug for now
 	}
 
 	halInstance, err := b.backend.CreateInstance(desc)
@@ -81,13 +88,44 @@ func (b *Backend) RequestAdapter(instance types.Instance, opts *types.AdapterOpt
 		return 0, fmt.Errorf("gpu: no adapters found")
 	}
 
-	// Pick first adapter for now
-	// TODO: Support power preference from opts
+	// Sort adapters based on power preference (matches wgpu-core behavior)
+	preferIntegrated := opts != nil && opts.PowerPreference == gputypes.PowerPreferenceLowPower
+	sort.SliceStable(adapters, func(i, j int) bool {
+		return adapterOrder(adapters[i].Info.DeviceType, preferIntegrated) <
+			adapterOrder(adapters[j].Info.DeviceType, preferIntegrated)
+	})
+
+	// Pick best adapter (first after sorting)
 	exposed := adapters[0]
 
 	// Register and return handle
 	handle := b.registry.RegisterAdapter(exposed.Adapter)
 	return handle, nil
+}
+
+// adapterOrder returns the priority order for adapter selection.
+// Lower values = higher priority. Matches wgpu-core's request_adapter behavior.
+func adapterOrder(deviceType gputypes.DeviceType, preferIntegrated bool) int {
+	switch deviceType {
+	case gputypes.DeviceTypeDiscreteGPU:
+		if preferIntegrated {
+			return 2
+		}
+		return 1 // Best for high performance
+	case gputypes.DeviceTypeIntegratedGPU:
+		if preferIntegrated {
+			return 1 // Best for low power
+		}
+		return 2
+	case gputypes.DeviceTypeOther:
+		return 3 // Unknown (could be OpenGL)
+	case gputypes.DeviceTypeVirtualGPU:
+		return 4
+	case gputypes.DeviceTypeCPU:
+		return 5 // Software fallback (worst)
+	default:
+		return 6
+	}
 }
 
 // RequestDevice requests a GPU device.
@@ -150,7 +188,7 @@ func (b *Backend) ConfigureSurface(surface types.Surface, device types.Device, c
 		return
 	}
 
-	// Store surface → device mapping for Present()
+	// Store surface -> device mapping for Present()
 	b.registry.RegisterSurfaceDevice(surface, device)
 
 	// Convert config - all types are now gputypes aliases, no conversion needed
@@ -174,9 +212,13 @@ func (b *Backend) GetCurrentTexture(surface types.Surface) (types.SurfaceTexture
 		return types.SurfaceTexture{Status: types.SurfaceStatusError}, err
 	}
 
-	// Acquire texture (fence=nil for now)
+	// Acquire texture (non-blocking)
 	acquired, err := halSurface.AcquireTexture(nil)
 	if err != nil {
+		// Check for "not ready" - this means skip frame, not an error
+		if errors.Is(err, hal.ErrNotReady) {
+			return types.SurfaceTexture{Status: types.SurfaceStatusTimeout}, nil
+		}
 		// Map HAL errors to surface status
 		return types.SurfaceTexture{Status: types.SurfaceStatusError}, err
 	}
@@ -184,12 +226,13 @@ func (b *Backend) GetCurrentTexture(surface types.Surface) (types.SurfaceTexture
 	// Store the SurfaceTexture for Present() to use later
 	b.registry.SetCurrentSurfaceTexture(surface, acquired.Texture)
 
-	// Register texture and return
+	// Get device for this surface (stored in ConfigureSurface)
 	device, err := b.registry.GetDeviceForSurface(surface)
 	if err != nil {
 		return types.SurfaceTexture{Status: types.SurfaceStatusError}, err
 	}
 
+	// Register texture WITH device (required for CreateTextureView)
 	textureHandle := b.registry.RegisterTextureForDevice(acquired.Texture, device)
 
 	return types.SurfaceTexture{
@@ -275,21 +318,30 @@ func (b *Backend) CreateRenderPipeline(device types.Device, desc *types.RenderPi
 		return 0, err
 	}
 
+	// Get pipeline layout if provided
+	var halLayout hal.PipelineLayout
+	if desc.Layout != 0 {
+		halLayout, err = b.registry.GetPipelineLayout(desc.Layout)
+		if err != nil {
+			return 0, fmt.Errorf("gpu: invalid pipeline layout: %w", err)
+		}
+	}
+
 	// Build HAL descriptor - types are now gputypes aliases, no conversion needed
 	halDesc := &hal.RenderPipelineDescriptor{
 		Label:  desc.Label,
-		Layout: nil, // Auto layout
+		Layout: halLayout,
 		Vertex: hal.VertexState{
 			Module:     vertexShader,
 			EntryPoint: desc.VertexEntryPoint,
-			Buffers:    nil, // No vertex buffers for triangle
+			Buffers:    nil, // No vertex buffers for fullscreen quad
 		},
 		Primitive: gputypes.PrimitiveState{
 			Topology:  desc.Topology,
 			FrontFace: desc.FrontFace,
 			CullMode:  desc.CullMode,
 		},
-		DepthStencil: nil, // No depth/stencil for triangle
+		DepthStencil: nil, // No depth/stencil
 		Multisample:  gputypes.MultisampleState{Count: 1, Mask: 0xFFFFFFFF},
 		Fragment: &hal.FragmentState{
 			Module:     fragmentShader,
@@ -297,7 +349,7 @@ func (b *Backend) CreateRenderPipeline(device types.Device, desc *types.RenderPi
 			Targets: []gputypes.ColorTargetState{
 				{
 					Format:    desc.TargetFormat,
-					Blend:     nil, // No blending for now
+					Blend:     desc.Blend,
 					WriteMask: gputypes.ColorWriteMaskAll,
 				},
 			},
@@ -329,7 +381,14 @@ func (b *Backend) CreateCommandEncoder(device types.Device) types.CommandEncoder
 		return 0
 	}
 
-	handle := b.registry.RegisterCommandEncoder(encoder)
+	// IMPORTANT: HAL requires BeginEncoding before any commands can be recorded.
+	// This must be called before BeginRenderPass, BeginComputePass, etc.
+	if err := encoder.BeginEncoding("frame"); err != nil {
+		return 0
+	}
+
+	// Register encoder with device so we can free the command buffer to correct pool
+	handle := b.registry.RegisterCommandEncoderForDevice(encoder, device)
 	return handle
 }
 
@@ -343,8 +402,8 @@ func (b *Backend) BeginRenderPass(encoder types.CommandEncoder, desc *types.Rend
 	// Convert color attachments - types are gputypes aliases now
 	colorAttachments := make([]hal.RenderPassColorAttachment, 0, len(desc.ColorAttachments))
 	for _, ca := range desc.ColorAttachments {
-		view, err := b.registry.GetTextureView(ca.View)
-		if err != nil {
+		view, viewErr := b.registry.GetTextureView(ca.View)
+		if viewErr != nil {
 			continue
 		}
 
@@ -385,12 +444,16 @@ func (b *Backend) FinishEncoder(encoder types.CommandEncoder) types.CommandBuffe
 		return 0
 	}
 
+	// Get the device this encoder was created from (for proper command buffer freeing)
+	device := b.registry.GetCommandEncoderDevice(encoder)
+
 	cmdBuffer, err := halEncoder.EndEncoding()
 	if err != nil {
 		return 0
 	}
 
-	handle := b.registry.RegisterCommandBuffer(cmdBuffer)
+	// Register command buffer with device so it can be freed to correct pool
+	handle := b.registry.RegisterCommandBufferForDevice(cmdBuffer, device)
 	return handle
 }
 
@@ -408,50 +471,39 @@ func (b *Backend) Submit(queue types.Queue, commands types.CommandBuffer, fence 
 		return 0
 	}
 
-	// Attach drawable from current surface texture to command buffer (Metal requirement).
-	// The drawable must be scheduled for presentation before commit.
-	b.attachDrawableToCommandBuffer(halCmdBuffer)
+	// Get HAL fence if provided
+	var halFence hal.Fence
+	if fence != 0 {
+		halFence, _ = b.registry.GetFence(fence)
+	}
 
-	// TODO: Pass fence to HAL when fence support is implemented.
-	// For now, submit without fence signaling.
-	_ = halQueue.Submit([]hal.CommandBuffer{halCmdBuffer}, nil, 0)
+	// Platform-specific pre-submit hook (Metal drawable attachment)
+	platformPreSubmit(halCmdBuffer, b.registry)
+
+	// Submit with fence signaling
+	_ = halQueue.Submit([]hal.CommandBuffer{halCmdBuffer}, halFence, fenceValue)
 
 	return types.SubmissionIndex(fenceValue)
 }
 
 // GetFenceStatus returns true if the fence is signaled (non-blocking).
 func (b *Backend) GetFenceStatus(fence types.Fence) (bool, error) {
-	// TODO: Implement fence status using Metal events when available.
-	return true, nil // Always signaled for now
-}
-
-// attachDrawableToCommandBuffer attaches the current drawable to a command buffer.
-// This is required for Metal where presentDrawable: must be called before commit.
-func (b *Backend) attachDrawableToCommandBuffer(cmdBuffer hal.CommandBuffer) {
-	// Type-assert to Metal command buffer
-	metalCmdBuffer, ok := cmdBuffer.(*metal.CommandBuffer)
-	if !ok {
-		return // Not Metal backend
+	halFence, err := b.registry.GetFence(fence)
+	if err != nil {
+		return false, err
 	}
 
-	// Find any current surface texture and get its drawable.
-	// In practice, there's only one surface per frame.
-	surfaceTexture := b.registry.GetAnySurfaceTexture()
-	if surfaceTexture == nil {
-		return
+	deviceHandle, err := b.registry.GetFenceDevice(fence)
+	if err != nil {
+		return false, err
 	}
 
-	// Type-assert to Metal surface texture
-	metalSurfaceTex, ok := surfaceTexture.(*metal.SurfaceTexture)
-	if !ok {
-		return
+	halDevice, err := b.registry.GetDevice(deviceHandle)
+	if err != nil {
+		return false, err
 	}
 
-	// Get drawable using accessor and attach to command buffer
-	drawable := metalSurfaceTex.Drawable()
-	if drawable != 0 {
-		metalCmdBuffer.SetDrawable(drawable)
-	}
+	return halDevice.GetFenceStatus(halFence)
 }
 
 // SetPipeline sets the render pipeline.
@@ -479,18 +531,42 @@ func (b *Backend) Draw(pass types.RenderPass, vertexCount, instanceCount, firstV
 	halPass.Draw(vertexCount, instanceCount, firstVertex, firstInstance)
 }
 
-// --- Texture operations (stubs for now) ---
+// --- Texture operations ---
 
+// CreateTexture creates a GPU texture.
 func (b *Backend) CreateTexture(device types.Device, desc *types.TextureDescriptor) (types.Texture, error) {
-	return 0, gpu.ErrNotImplemented
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: invalid device: %w", err)
+	}
+
+	halDesc := &hal.TextureDescriptor{
+		Label:         desc.Label,
+		Size:          hal.Extent3D{Width: desc.Size.Width, Height: desc.Size.Height, DepthOrArrayLayers: desc.Size.DepthOrArrayLayers},
+		MipLevelCount: desc.MipLevelCount,
+		SampleCount:   desc.SampleCount,
+		Dimension:     desc.Dimension,
+		Format:        desc.Format,
+		Usage:         desc.Usage,
+	}
+
+	texture, err := halDevice.CreateTexture(halDesc)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: failed to create texture: %w", err)
+	}
+
+	handle := b.registry.RegisterTextureForDevice(texture, device)
+	return handle, nil
 }
 
+// CreateTextureView creates a texture view.
 func (b *Backend) CreateTextureView(texture types.Texture, desc *types.TextureViewDescriptor) types.TextureView {
 	halTexture, err := b.registry.GetTexture(texture)
 	if err != nil {
 		return 0
 	}
 
+	// Get device for this texture (stored in RegisterTextureForDevice)
 	deviceHandle, err := b.registry.GetDeviceForTexture(texture)
 	if err != nil {
 		return 0
@@ -524,102 +600,460 @@ func (b *Backend) CreateTextureView(texture types.Texture, desc *types.TextureVi
 	return handle
 }
 
+// WriteTexture writes data to a texture.
 func (b *Backend) WriteTexture(queue types.Queue, dst *types.ImageCopyTexture, data []byte, layout *types.ImageDataLayout, size *gputypes.Extent3D) {
-	// Not implemented yet
+	halQueue, err := b.registry.GetQueue(queue)
+	if err != nil {
+		return // Silent fail for now, matches wgpu behavior
+	}
+
+	halTexture, err := b.registry.GetTexture(dst.Texture)
+	if err != nil {
+		return
+	}
+
+	halDst := &hal.ImageCopyTexture{
+		Texture:  halTexture,
+		MipLevel: dst.MipLevel,
+		Origin:   hal.Origin3D{X: dst.Origin.X, Y: dst.Origin.Y, Z: dst.Origin.Z},
+		Aspect:   dst.Aspect,
+	}
+
+	halLayout := &hal.ImageDataLayout{
+		Offset:       layout.Offset,
+		BytesPerRow:  layout.BytesPerRow,
+		RowsPerImage: layout.RowsPerImage,
+	}
+
+	halSize := &hal.Extent3D{
+		Width:              size.Width,
+		Height:             size.Height,
+		DepthOrArrayLayers: size.DepthOrArrayLayers,
+	}
+
+	halQueue.WriteTexture(halDst, data, halLayout, halSize)
 }
 
+// CreateSampler creates a texture sampler.
 func (b *Backend) CreateSampler(device types.Device, desc *types.SamplerDescriptor) (types.Sampler, error) {
-	return 0, gpu.ErrNotImplemented
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: invalid device: %w", err)
+	}
+
+	halDesc := &hal.SamplerDescriptor{
+		Label:        desc.Label,
+		AddressModeU: desc.AddressModeU,
+		AddressModeV: desc.AddressModeV,
+		AddressModeW: desc.AddressModeW,
+		MagFilter:    desc.MagFilter,
+		MinFilter:    desc.MinFilter,
+		MipmapFilter: gputypes.FilterMode(desc.MipmapFilter),
+		LodMinClamp:  desc.LodMinClamp,
+		LodMaxClamp:  desc.LodMaxClamp,
+		Compare:      desc.Compare,
+		Anisotropy:   desc.MaxAnisotropy,
+	}
+
+	sampler, err := halDevice.CreateSampler(halDesc)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: failed to create sampler: %w", err)
+	}
+
+	handle := b.registry.RegisterSampler(sampler)
+	return handle, nil
 }
 
+// CreateBuffer creates a GPU buffer.
 func (b *Backend) CreateBuffer(device types.Device, desc *types.BufferDescriptor) (types.Buffer, error) {
-	return 0, gpu.ErrNotImplemented
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: invalid device: %w", err)
+	}
+
+	// For uniform/copy-dst buffers, we need mapped memory for WriteBuffer to work
+	// Native HAL doesn't have staging buffer support yet, so we use host-visible memory
+	mappedAtCreation := desc.MappedAtCreation
+	if desc.Usage&gputypes.BufferUsageCopyDst != 0 {
+		mappedAtCreation = true
+	}
+
+	halDesc := &hal.BufferDescriptor{
+		Label:            desc.Label,
+		Size:             desc.Size,
+		Usage:            desc.Usage,
+		MappedAtCreation: mappedAtCreation,
+	}
+
+	buffer, err := halDevice.CreateBuffer(halDesc)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: failed to create buffer: %w", err)
+	}
+
+	handle := b.registry.RegisterBuffer(buffer)
+	b.bufferSizes[handle] = desc.Size
+	return handle, nil
 }
 
+// WriteBuffer writes data to a buffer.
 func (b *Backend) WriteBuffer(queue types.Queue, buffer types.Buffer, offset uint64, data []byte) {
-	// Not implemented yet
+	halQueue, err := b.registry.GetQueue(queue)
+	if err != nil {
+		return // Silent fail, matching Rust backend behavior
+	}
+
+	halBuffer, err := b.registry.GetBuffer(buffer)
+	if err != nil {
+		return
+	}
+
+	halQueue.WriteBuffer(halBuffer, offset, data)
 }
 
-func (b *Backend) CopyBufferToBuffer(types.CommandEncoder, types.Buffer, uint64, types.Buffer, uint64, uint64) {
+// CopyBufferToBuffer records a buffer-to-buffer copy command.
+func (b *Backend) CopyBufferToBuffer(encoder types.CommandEncoder, src types.Buffer, srcOffset uint64, dst types.Buffer, dstOffset, size uint64) {
+	halEncoder, err := b.registry.GetCommandEncoder(encoder)
+	if err != nil {
+		return
+	}
+	halSrc, err := b.registry.GetBuffer(src)
+	if err != nil {
+		return
+	}
+	halDst, err := b.registry.GetBuffer(dst)
+	if err != nil {
+		return
+	}
+	halEncoder.CopyBufferToBuffer(halSrc, halDst, []hal.BufferCopy{
+		{SrcOffset: srcOffset, DstOffset: dstOffset, Size: size},
+	})
 }
 
+// CreateBindGroupLayout creates a bind group layout.
 func (b *Backend) CreateBindGroupLayout(device types.Device, desc *types.BindGroupLayoutDescriptor) (types.BindGroupLayout, error) {
-	return 0, gpu.ErrNotImplemented
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: invalid device: %w", err)
+	}
+
+	// Convert entries to HAL format
+	halEntries := make([]gputypes.BindGroupLayoutEntry, len(desc.Entries))
+	for i, entry := range desc.Entries {
+		halEntries[i] = gputypes.BindGroupLayoutEntry{
+			Binding:    entry.Binding,
+			Visibility: entry.Visibility,
+			Buffer:     entry.Buffer,
+			Sampler:    entry.Sampler,
+			Texture:    entry.Texture,
+		}
+	}
+
+	halDesc := &hal.BindGroupLayoutDescriptor{
+		Label:   desc.Label,
+		Entries: halEntries,
+	}
+
+	layout, err := halDevice.CreateBindGroupLayout(halDesc)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: failed to create bind group layout: %w", err)
+	}
+
+	handle := b.registry.RegisterBindGroupLayout(layout)
+	return handle, nil
 }
 
+// CreateBindGroup creates a bind group.
 func (b *Backend) CreateBindGroup(device types.Device, desc *types.BindGroupDescriptor) (types.BindGroup, error) {
-	return 0, gpu.ErrNotImplemented
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: invalid device: %w", err)
+	}
+
+	halLayout, err := b.registry.GetBindGroupLayout(desc.Layout)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: invalid bind group layout: %w", err)
+	}
+
+	// Convert entries - need to resolve handles to gpu handles
+	halEntries := make([]gputypes.BindGroupEntry, len(desc.Entries))
+	for i, entry := range desc.Entries {
+		halEntries[i] = gputypes.BindGroupEntry{
+			Binding: entry.Binding,
+		}
+
+		// Determine which resource is set and convert using NativeHandle
+		switch {
+		case entry.Buffer != 0:
+			halBuffer, bufErr := b.registry.GetBuffer(entry.Buffer)
+			if bufErr != nil {
+				return 0, fmt.Errorf("gpu: invalid buffer in bind group entry %d: %w", i, bufErr)
+			}
+			halEntries[i].Resource = gputypes.BufferBinding{
+				Buffer: halBuffer.NativeHandle(),
+				Offset: entry.Offset,
+				Size:   entry.Size,
+			}
+		case entry.Sampler != 0:
+			halSampler, sampErr := b.registry.GetSampler(entry.Sampler)
+			if sampErr != nil {
+				return 0, fmt.Errorf("gpu: invalid sampler in bind group entry %d: %w", i, sampErr)
+			}
+			halEntries[i].Resource = gputypes.SamplerBinding{
+				Sampler: halSampler.NativeHandle(),
+			}
+		case entry.TextureView != 0:
+			halView, viewErr := b.registry.GetTextureView(entry.TextureView)
+			if viewErr != nil {
+				return 0, fmt.Errorf("gpu: invalid texture view in bind group entry %d: %w", i, viewErr)
+			}
+			halEntries[i].Resource = gputypes.TextureViewBinding{
+				TextureView: halView.NativeHandle(),
+			}
+		default:
+			return 0, fmt.Errorf("gpu: bind group entry %d has no resource", i)
+		}
+	}
+
+	halDesc := &hal.BindGroupDescriptor{
+		Label:   desc.Label,
+		Layout:  halLayout,
+		Entries: halEntries,
+	}
+
+	group, err := halDevice.CreateBindGroup(halDesc)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: failed to create bind group: %w", err)
+	}
+
+	handle := b.registry.RegisterBindGroup(group)
+	return handle, nil
 }
 
+// CreatePipelineLayout creates a pipeline layout.
 func (b *Backend) CreatePipelineLayout(device types.Device, desc *types.PipelineLayoutDescriptor) (types.PipelineLayout, error) {
-	return 0, gpu.ErrNotImplemented
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: invalid device: %w", err)
+	}
+
+	// Convert bind group layouts
+	halLayouts := make([]hal.BindGroupLayout, len(desc.BindGroupLayouts))
+	for i, layout := range desc.BindGroupLayouts {
+		halLayout, layoutErr := b.registry.GetBindGroupLayout(layout)
+		if layoutErr != nil {
+			return 0, fmt.Errorf("gpu: invalid bind group layout at index %d: %w", i, layoutErr)
+		}
+		halLayouts[i] = halLayout
+	}
+
+	halDesc := &hal.PipelineLayoutDescriptor{
+		Label:            desc.Label,
+		BindGroupLayouts: halLayouts,
+	}
+
+	layout, err := halDevice.CreatePipelineLayout(halDesc)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: failed to create pipeline layout: %w", err)
+	}
+
+	handle := b.registry.RegisterPipelineLayout(layout)
+	return handle, nil
 }
 
+// SetBindGroup sets a bind group for a render pass.
 func (b *Backend) SetBindGroup(pass types.RenderPass, index uint32, bindGroup types.BindGroup, dynamicOffsets []uint32) {
-	// Not implemented yet
+	halPass, err := b.registry.GetRenderPass(pass)
+	if err != nil {
+		return
+	}
+
+	halGroup, err := b.registry.GetBindGroup(bindGroup)
+	if err != nil {
+		return
+	}
+
+	halPass.SetBindGroup(index, halGroup, dynamicOffsets)
 }
 
+// SetVertexBuffer sets a vertex buffer for a render pass.
 func (b *Backend) SetVertexBuffer(pass types.RenderPass, slot uint32, buffer types.Buffer, offset, size uint64) {
 	// Not implemented yet
 }
 
+// SetIndexBuffer sets an index buffer for a render pass.
 func (b *Backend) SetIndexBuffer(pass types.RenderPass, buffer types.Buffer, format gputypes.IndexFormat, offset, size uint64) {
 	// Not implemented yet
 }
 
+// DrawIndexed issues an indexed draw call.
 func (b *Backend) DrawIndexed(pass types.RenderPass, indexCount, instanceCount, firstIndex uint32, baseVertex int32, firstInstance uint32) {
 	// Not implemented yet
 }
 
-// --- Compute shader operations (stubs) ---
+// --- Compute shader operations ---
 
 // CreateShaderModuleSPIRV creates a shader module from SPIR-V bytecode.
 func (b *Backend) CreateShaderModuleSPIRV(device types.Device, spirv []uint32) (types.ShaderModule, error) {
-	return 0, gpu.ErrNotImplemented
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return 0, err
+	}
+
+	desc := &hal.ShaderModuleDescriptor{
+		Label:  "shader-spirv",
+		Source: hal.ShaderSource{SPIRV: spirv},
+	}
+
+	module, err := halDevice.CreateShaderModule(desc)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: failed to create SPIR-V shader module: %w", err)
+	}
+
+	handle := b.registry.RegisterShaderModule(module)
+	return handle, nil
 }
 
 // CreateComputePipeline creates a compute pipeline.
 func (b *Backend) CreateComputePipeline(device types.Device, desc *types.ComputePipelineDescriptor) (types.ComputePipeline, error) {
-	return 0, gpu.ErrNotImplemented
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return 0, err
+	}
+
+	// Get shader module
+	halModule, err := b.registry.GetShaderModule(desc.Module)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: invalid compute shader module: %w", err)
+	}
+
+	// Get pipeline layout if provided
+	var halLayout hal.PipelineLayout
+	if desc.Layout != 0 {
+		halLayout, err = b.registry.GetPipelineLayout(desc.Layout)
+		if err != nil {
+			return 0, fmt.Errorf("gpu: invalid pipeline layout: %w", err)
+		}
+	}
+
+	halDesc := &hal.ComputePipelineDescriptor{
+		Label:  desc.Label,
+		Layout: halLayout,
+		Compute: hal.ComputeState{
+			Module:     halModule,
+			EntryPoint: desc.EntryPoint,
+		},
+	}
+
+	pipeline, err := halDevice.CreateComputePipeline(halDesc)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: failed to create compute pipeline: %w", err)
+	}
+
+	handle := b.registry.RegisterComputePipeline(pipeline)
+	return handle, nil
 }
 
 // BeginComputePass begins a compute pass.
 func (b *Backend) BeginComputePass(encoder types.CommandEncoder) types.ComputePass {
-	return 0
+	halEncoder, err := b.registry.GetCommandEncoder(encoder)
+	if err != nil {
+		return 0
+	}
+
+	halDesc := &hal.ComputePassDescriptor{
+		Label: "compute_pass",
+	}
+
+	pass := halEncoder.BeginComputePass(halDesc)
+
+	handle := b.registry.RegisterComputePass(pass)
+	return handle
 }
 
 // EndComputePass ends a compute pass.
 func (b *Backend) EndComputePass(pass types.ComputePass) {
-	// Not implemented yet
+	halPass, err := b.registry.GetComputePass(pass)
+	if err != nil {
+		return
+	}
+
+	halPass.End()
 }
 
 // SetComputePipeline sets the compute pipeline for a compute pass.
 func (b *Backend) SetComputePipeline(pass types.ComputePass, pipeline types.ComputePipeline) {
-	// Not implemented yet
+	halPass, err := b.registry.GetComputePass(pass)
+	if err != nil {
+		return
+	}
+
+	halPipeline, err := b.registry.GetComputePipeline(pipeline)
+	if err != nil {
+		return
+	}
+
+	halPass.SetPipeline(halPipeline)
 }
 
 // SetComputeBindGroup sets a bind group for a compute pass.
 func (b *Backend) SetComputeBindGroup(pass types.ComputePass, index uint32, bindGroup types.BindGroup, dynamicOffsets []uint32) {
-	// Not implemented yet
+	halPass, err := b.registry.GetComputePass(pass)
+	if err != nil {
+		return
+	}
+
+	halGroup, err := b.registry.GetBindGroup(bindGroup)
+	if err != nil {
+		return
+	}
+
+	halPass.SetBindGroup(index, halGroup, dynamicOffsets)
 }
 
 // DispatchWorkgroups dispatches compute work.
 func (b *Backend) DispatchWorkgroups(pass types.ComputePass, x, y, z uint32) {
-	// Not implemented yet
+	halPass, err := b.registry.GetComputePass(pass)
+	if err != nil {
+		return
+	}
+
+	halPass.Dispatch(x, y, z)
 }
 
-// MapBufferRead maps a buffer for reading and returns its contents.
+// MapBufferRead reads a buffer's contents from the GPU.
+// The buffer must have been created with BufferUsageMapRead | BufferUsageCopyDst.
 func (b *Backend) MapBufferRead(buffer types.Buffer) ([]byte, error) {
-	return nil, gpu.ErrNotImplemented
+	halBuffer, err := b.registry.GetBuffer(buffer)
+	if err != nil {
+		return nil, fmt.Errorf("gpu: invalid buffer: %w", err)
+	}
+
+	size, ok := b.bufferSizes[buffer]
+	if !ok {
+		return nil, fmt.Errorf("gpu: unknown buffer size for handle %d", buffer)
+	}
+
+	halQueue := b.registry.GetAnyQueue()
+	if halQueue == nil {
+		return nil, fmt.Errorf("gpu: no queue available for ReadBuffer")
+	}
+
+	data := make([]byte, size)
+	if err := halQueue.ReadBuffer(halBuffer, 0, data); err != nil {
+		return nil, fmt.Errorf("gpu: ReadBuffer failed: %w", err)
+	}
+	return data, nil
 }
 
-// UnmapBuffer unmaps a previously mapped buffer.
+// UnmapBuffer is a no-op for the native backend.
+// ReadBuffer returns a copy of the data, so no unmapping is needed.
 func (b *Backend) UnmapBuffer(buffer types.Buffer) {
-	// Not implemented yet
+	// No-op: ReadBuffer copies data, no persistent mapping to release.
 }
 
 // --- Resource release ---
 
+// ReleaseTexture releases a texture.
 func (b *Backend) ReleaseTexture(texture types.Texture) {
 	halTexture, err := b.registry.GetTexture(texture)
 	if err == nil && halTexture != nil {
@@ -628,6 +1062,7 @@ func (b *Backend) ReleaseTexture(texture types.Texture) {
 	b.registry.UnregisterTexture(texture)
 }
 
+// ReleaseTextureView releases a texture view.
 func (b *Backend) ReleaseTextureView(view types.TextureView) {
 	halView, err := b.registry.GetTextureView(view)
 	if err == nil && halView != nil {
@@ -636,6 +1071,7 @@ func (b *Backend) ReleaseTextureView(view types.TextureView) {
 	b.registry.UnregisterTextureView(view)
 }
 
+// ReleaseSampler releases a sampler.
 func (b *Backend) ReleaseSampler(sampler types.Sampler) {
 	halSampler, err := b.registry.GetSampler(sampler)
 	if err == nil && halSampler != nil {
@@ -644,14 +1080,17 @@ func (b *Backend) ReleaseSampler(sampler types.Sampler) {
 	b.registry.UnregisterSampler(sampler)
 }
 
+// ReleaseBuffer releases a buffer.
 func (b *Backend) ReleaseBuffer(buffer types.Buffer) {
 	halBuffer, err := b.registry.GetBuffer(buffer)
 	if err == nil && halBuffer != nil {
 		halBuffer.Destroy()
 	}
 	b.registry.UnregisterBuffer(buffer)
+	delete(b.bufferSizes, buffer)
 }
 
+// ReleaseBindGroupLayout releases a bind group layout.
 func (b *Backend) ReleaseBindGroupLayout(layout types.BindGroupLayout) {
 	halLayout, err := b.registry.GetBindGroupLayout(layout)
 	if err == nil && halLayout != nil {
@@ -660,6 +1099,7 @@ func (b *Backend) ReleaseBindGroupLayout(layout types.BindGroupLayout) {
 	b.registry.UnregisterBindGroupLayout(layout)
 }
 
+// ReleaseBindGroup releases a bind group.
 func (b *Backend) ReleaseBindGroup(group types.BindGroup) {
 	halGroup, err := b.registry.GetBindGroup(group)
 	if err == nil && halGroup != nil {
@@ -668,6 +1108,7 @@ func (b *Backend) ReleaseBindGroup(group types.BindGroup) {
 	b.registry.UnregisterBindGroup(group)
 }
 
+// ReleasePipelineLayout releases a pipeline layout.
 func (b *Backend) ReleasePipelineLayout(layout types.PipelineLayout) {
 	halLayout, err := b.registry.GetPipelineLayout(layout)
 	if err == nil && halLayout != nil {
@@ -676,20 +1117,30 @@ func (b *Backend) ReleasePipelineLayout(layout types.PipelineLayout) {
 	b.registry.UnregisterPipelineLayout(layout)
 }
 
+// ReleaseCommandBuffer releases a command buffer.
 func (b *Backend) ReleaseCommandBuffer(buffer types.CommandBuffer) {
 	halBuffer, err := b.registry.GetCommandBuffer(buffer)
 	if err == nil && halBuffer != nil {
-		halBuffer.Destroy()
+		// Get device to free command buffer back to pool
+		deviceHandle := b.registry.GetCommandBufferDevice(buffer)
+		if deviceHandle != 0 {
+			halDevice, devErr := b.registry.GetDevice(deviceHandle)
+			if devErr == nil {
+				halDevice.FreeCommandBuffer(halBuffer)
+			}
+		}
 	}
 	b.registry.UnregisterCommandBuffer(buffer)
 }
 
+// ReleaseCommandEncoder releases a command encoder.
 func (b *Backend) ReleaseCommandEncoder(encoder types.CommandEncoder) {
 	// Command encoders don't have Destroy in HAL - they're consumed when EndEncoding() is called.
 	// We just unregister the handle from the registry.
 	b.registry.UnregisterCommandEncoder(encoder)
 }
 
+// ReleaseRenderPass releases a render pass.
 func (b *Backend) ReleaseRenderPass(pass types.RenderPass) {
 	// Render passes are ended, not destroyed
 	b.registry.UnregisterRenderPass(pass)
@@ -697,12 +1148,17 @@ func (b *Backend) ReleaseRenderPass(pass types.RenderPass) {
 
 // ReleaseComputePipeline releases a compute pipeline.
 func (b *Backend) ReleaseComputePipeline(pipeline types.ComputePipeline) {
-	// Not implemented yet
+	halPipeline, err := b.registry.GetComputePipeline(pipeline)
+	if err == nil && halPipeline != nil {
+		halPipeline.Destroy()
+	}
+	b.registry.UnregisterComputePipeline(pipeline)
 }
 
 // ReleaseComputePass releases a compute pass.
 func (b *Backend) ReleaseComputePass(pass types.ComputePass) {
-	// Not implemented yet
+	// Compute passes are ended, not destroyed
+	b.registry.UnregisterComputePass(pass)
 }
 
 // ReleaseShaderModule releases a shader module.
@@ -715,34 +1171,88 @@ func (b *Backend) ReleaseShaderModule(module types.ShaderModule) {
 }
 
 // ResetCommandPool resets the command pool to reclaim command buffer memory.
-// Metal uses automatic command buffer management, so this is a no-op.
+// This waits for GPU to finish all operations first (blocking).
 func (b *Backend) ResetCommandPool(device types.Device) {
-	// Metal manages command buffers automatically through MTLCommandQueue.
-	// No explicit reset needed.
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return
+	}
+
+	// Type assert to access ResetCommandPool (available on backends that support it)
+	type commandPoolResetter interface {
+		WaitIdle() error
+		ResetCommandPool() error
+	}
+	if resetter, ok := halDevice.(commandPoolResetter); ok {
+		// Wait for GPU to finish using command buffers
+		_ = resetter.WaitIdle()
+		// Reset the pool to reclaim memory
+		_ = resetter.ResetCommandPool()
+	}
 }
 
 // CreateFence creates a new fence in the unsignaled state.
-// Metal uses MTLEvent or MTLSharedEvent for synchronization.
 func (b *Backend) CreateFence(device types.Device) (types.Fence, error) {
-	// TODO: Implement fence creation using Metal events when available
-	return 0, gpu.ErrNotImplemented
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return 0, fmt.Errorf("gpu: invalid device: %w", err)
+	}
+
+	halFence, err := halDevice.CreateFence()
+	if err != nil {
+		return 0, fmt.Errorf("gpu: failed to create fence: %w", err)
+	}
+
+	handle := b.registry.RegisterFence(halFence, device)
+	return handle, nil
 }
 
 // WaitFence waits for a fence to be signaled.
 func (b *Backend) WaitFence(device types.Device, fence types.Fence, timeout uint64) (bool, error) {
-	// TODO: Implement fence waiting using Metal events when available
-	return true, nil // Always "signaled" for now
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return false, fmt.Errorf("gpu: invalid device: %w", err)
+	}
+
+	halFence, err := b.registry.GetFence(fence)
+	if err != nil {
+		return false, fmt.Errorf("gpu: invalid fence: %w", err)
+	}
+
+	// Convert timeout from nanoseconds to time.Duration.
+	// Max int64 is ~292 years in nanoseconds - any practical timeout is safe.
+	return halDevice.Wait(halFence, 0, time.Duration(timeout)) //nolint:gosec // G115: practical timeouts won't overflow
 }
 
 // ResetFence resets a fence to the unsignaled state.
 func (b *Backend) ResetFence(device types.Device, fence types.Fence) error {
-	// TODO: Implement fence reset using Metal events when available
-	return nil
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return fmt.Errorf("gpu: invalid device: %w", err)
+	}
+
+	halFence, err := b.registry.GetFence(fence)
+	if err != nil {
+		return fmt.Errorf("gpu: invalid fence: %w", err)
+	}
+
+	return halDevice.ResetFence(halFence)
 }
 
 // DestroyFence destroys a fence.
 func (b *Backend) DestroyFence(device types.Device, fence types.Fence) {
-	// TODO: Implement fence destruction using Metal events when available
+	halDevice, err := b.registry.GetDevice(device)
+	if err != nil {
+		return
+	}
+
+	halFence, err := b.registry.GetFence(fence)
+	if err != nil {
+		return
+	}
+
+	halDevice.DestroyFence(halFence)
+	b.registry.UnregisterFence(fence)
 }
 
 // GetHalDevice returns the underlying HAL device for the given handle.
